added filtering logic to database reads
In the the data model, I mimicked what the router is currently doing.
Still trying to get the controller to call this function and pass the data back to the router, keep hitting error: "could not read length of undefined" so somehow I am not getting any data back from the database. I think it is something todo with the async nature of mongoose data reads and the fact that the webserver is responding faster than the database can. since there's no logic telling pug to wait on the data object it throws an error.