attempt fix at rendering in index.pug
whenever we are reading from the database, for some reason the data is not making it back to the pug template. console.log shows the data is being read, index.pug renders with the table written out for the data but none of the data will make it to the table...it remains blank.