updated readVATSIM()
finally figured out that with a callback method I can offload the weight of the database lookup to the CRUD class and maintain clean arch design. This works as well for passing the info back to the pug template, previously if the call was in the CRUD class the template would render without the data because reading from database is slow.  This logic is largely based off a dated article: https://webapplog.com/url-parameters-and-routing-in-express-js/